requirements.txt
Django>=3.2.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.2.0
python-dotenv>=0.19.0
requests==2.31.0
psycopg2>=2.9.0